Fix signed-unsigned conversion warnings introduced in 9258ac6

+++ src/libty/monitor.c
@@ -152,7 +152,7 @@ static int add_missing_board(ty_board *board)
     board = ty_list_get_first(&monitor->missing_boards, ty_board, missing_node);
 
     return ty_timer_set(monitor->timer,
-                        ty_adjust_timeout(monitor->drop_delay, board->missing_since),
+                        ty_adjust_timeout((int)monitor->drop_delay, board->missing_since),
                         TY_TIMER_ONESHOT);
 }
 
@@ -577,7 +577,7 @@ int ty_monitor_refresh(ty_monitor *monitor)
             ty_board *board = ty_container_of(cur, ty_board, missing_node);
             int timeout;
 
-            timeout = ty_adjust_timeout(monitor->drop_delay, board->missing_since);
+            timeout = ty_adjust_timeout((int)monitor->drop_delay, board->missing_since);
             if (timeout) {
                 r = ty_timer_set(monitor->timer, timeout, TY_TIMER_ONESHOT);
                 if (r < 0)
